feature/fix-type []: 반환값 통일 int형으로

See merge request s09-webmobile1-sub2/S09P12A508!123

+++ validate_server/JavaValidation/problemsolvingjava/src/main/java/com/problemsolvingjava/validate/controller/CodeExecutionController.java
@@ -206,7 +206,7 @@ public class CodeExecutionController {
         System.out.println(maxElapsedTime);
 
         compileResultDto.setResult(1);
-        compileResultDto.setExecutionTime(maxElapsedTime / 1_000_000);
+        compileResultDto.setExecutionTime((int)(maxElapsedTime / 1_000_000));
         compileResultDto.setMemoryUsage((int) (maxMemoryUsed.get()/1024));
         compileResultDto.setOutput(outputList);
         return new ResponseEntity<>(compileResultDto, HttpStatus.OK);
@@ -370,7 +370,7 @@ public class CodeExecutionController {
         }
 
         validationResultDto.setResult(1);
-        validationResultDto.setExecutionTime(maxElapsedTime / 1_000_000);
+        validationResultDto.setExecutionTime((int)(maxElapsedTime / 1_000_000));
         validationResultDto.setMemoryUsage((int) (maxMemoryUsed.get()/1024));
         return new ResponseEntity<>(validationResultDto, HttpStatus.OK);
     }
